Move files under src/
